+++ src/lang/js/definitions.rs
@@ -1,16 +1,75 @@
 // devela::lang:js::namespace
 //
-//! Defines the [`Js`] namespace and other items.
+//! Defines the [`Js`] namespace and related items.
 //
 // TOC
 // - struct Js
+// - enum JsEvent
 // - enum JsPermission
 // - enum JsPermissionState
 
 use crate::{TAG_EXPERIMENTAL, TAG_NON_STANDARD};
 
 /// A Javascript namespace.
+///
+/// # Methods
+/// - core APis
+///   - [console](#web-api-console)
+///   - [events](#web-api-events)
+///   - [history](#web-api-history--navigation)
+///   - [permissions](#web-api-permissions)
+/// - extended APis
+///   - media & graphics
+///     - [canvas](#web-api-canvas)
+//   - system & hardware
+///   - performance & optimization
+//     - time
+//   - advanced & experimental
 pub struct Js;
+
+/// # Web API Events
+///
+/// - <https://developer.mozilla.org/en-US/docs/Web/API/Event>
+/// - <https://developer.mozilla.org/en-US/docs/Web/API/EventTarget>
+#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
+#[repr(u8)]
+pub enum JsEvent {
+    /// Fires when an element is clicked.
+    Click,
+
+    /// Fires when a key is pressed down.
+    KeyDown,
+
+    /// Fires when a key is released.
+    KeyUp,
+
+    /// Fires when the mouse moves over an element.
+    MouseMove,
+
+    /// Fires when the mouse button is pressed down.
+    MouseDown,
+
+    /// Fires when the mouse button is released.
+    MouseUp,
+
+    /// Fires when the window is resized.
+    Resize,
+}
+impl JsEvent {
+    /// Returns the event name as a string.
+    pub fn as_str(self) -> &'static str {
+        use JsEvent as E;
+        match self {
+            E::Click => "click",
+            E::KeyDown => "keydown",
+            E::KeyUp => "keyup",
+            E::MouseMove => "mousemove",
+            E::MouseDown => "mousedown",
+            E::MouseUp => "mouseup",
+            E::Resize => "resize",
+        }
+    }
+}
 
 /// # Web API permissions
 ///
@@ -75,7 +134,6 @@ pub enum JsPermission {
     /// Allows a site to access storage without top-level navigation.
     TopLevelStorageAccess,
 }
-
 impl JsPermission {
     /// Returns the permission name as a string.
     pub fn as_str(self) -> &'static str {